updates  the components  and  fix  bugs

+++ src/components/ModelCard.js
@@ -1,21 +1,25 @@
-import React from 'react';
-// import  EditModel from './EditModel'
+import React, {useState} from 'react';
 
-export default function ModelCard({modelThumbnail, modelTitle, modelDescription, modelReleaseYear, modelCity, models, setModels}){
-     
+export default function ModelCard({modelThumbnail, modelTitle, modelDescription, modelReleaseYear, modelCity}){
+
+ const[models, setModels] = useState([])
      
     function handleDeleteClick(id){
-        fetch(`http://localhost:4000/models/${id}`,{
+      if (window.confirm("Do you want to delete this post?"))
+
+        fetch(`http://localhost:9292/models/${id}`,{
           method:"DELETE",
         })
         .then(res => res.json())
         .then(() => {
-          const deleteModel = models.filter((models) => models.id !== id)
-          setModels(deleteModel)
+          const deleteModel = models.filter(
+            (model) => model.id !== id
+            )
+            setModels(...models, deleteModel)
         })
       
     }
-
+// console.log(models)
     return(
         <div className="col m-2">
         <div className='card' style={{width : 25 + 'rem'}} > 
@@ -25,10 +29,10 @@ export default function ModelCard({modelThumbnail, modelTitle, modelDescription,
             ReleaseYear: &nbsp;{modelReleaseYear}<br></br>
             Description: &nbsp;{modelDescription}<br></br>
             City:   &nbsp;{modelCity}
-
-            <button className="remove" onClick={(e) =>handleDeleteClick(models.id)}> <span role="img" aria-label="delete">
+            <button className="remove" onClick={()=>handleDeleteClick(models.id)}> <span role="img" aria-label="delete">
               🗑
             </span>Delete</button>
+            
         </div>
         </div>
